+++ server/component/nav/tpl_nav.php
@@ -15,6 +15,7 @@
             </ul>
             <ul class="navbar-nav navbar-right">
                 <?php
+                    $this->output_nav_contact();
                     if($_SESSION['logged_in'])
                         $this->output_nav_menu('profile', $profile_title, $profile_children, true);
                     else